Flask-app/requirement.txt
Flask
slackclient
slackeventsapi
plotly==4.9.0
kaleido